ZWinSlider is now templated
Shadow rendering in textbox is beautiful now.

+++ SandboxApps/TextTestWin.h
@@ -37,7 +37,7 @@ private:
 
     ZGUI::TextBox mTextBox;
     bool            mbViewShadow;
-    int64_t         mShadowSpread;
+    float           mShadowSpread;
 
     ZGUI::Palette mPalette;
 };
